# src/main/resources/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Movie Ticket Booking API
  description: 'Browse screenings, reserve seats, complete bookings, cancel bookings, and administer the
    local catalog. This document is the source of truth: request/response models and the controller interfaces
    are generated from it at build time.'
  version: v1
servers:
- url: http://localhost:8080
  description: Local development
...
paths:
  /admin/api/v1/theaters/{id}:
    get:
      tags:
      - Admin catalog
      operationId: getTheater
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TheaterResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
      - basicAuth: []
    put:
      tags:
      - Admin catalog
      operationId: updateTheater
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TheaterRequest'
            example:
              name: Central Cinema
              cityId: 00000000-0000-4000-8000-000000000001
              address: Baner, Pune
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TheaterResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
    delete:
      tags:
      - Admin catalog
      operationId: deactivateTheater
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: No Content
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
  /admin/api/v1/seats/{id}:
    get:
      tags:
      - Admin catalog
      operationId: getSeat
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
      - basicAuth: []
    put:
      tags:
      - Admin catalog
      operationId: updateSeat
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeatRequest'
            example:
              rowLabel: A
              category: REGULAR
              seatNumber: 1
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
    delete:
      tags:
      - Admin catalog
      operationId: deactivateSeat
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: No Content
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
...
  /admin/api/v1/movies/{id}:
    get:
      tags:
      - Admin catalog
      operationId: getMovie
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
      - basicAuth: []
    put:
      tags:
      - Admin catalog
      operationId: updateMovie
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieRequest'
            example:
              durationMinutes: 120
              title: The Last Commit
              language: English
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
    delete:
      tags:
      - Admin catalog
      operationId: deactivateMovie
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: No Content
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
...
  /admin/api/v1/cities/{id}:
    get:
      tags:
      - Admin catalog
      operationId: getCity
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
      - basicAuth: []
    put:
      tags:
      - Admin catalog
      operationId: updateCity
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CityRequest'
            example:
              name: Pune
              timeZone: Asia/Kolkata
              country: India
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
    delete:
      tags:
      - Admin catalog
      operationId: deactivateCity
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: No Content
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
  /admin/api/v1/auditoriums/{id}:
    get:
      tags:
      - Admin catalog
      operationId: getAuditorium
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditoriumResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
      - basicAuth: []
    put:
      tags:
      - Admin catalog
      operationId: updateAuditorium
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditoriumRequest'
            example:
              name: Screen 1
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditoriumResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
    delete:
      tags:
      - Admin catalog
      operationId: deactivateAuditorium
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: No Content
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
...
  /admin/api/v1/theaters:
    get:
      tags:
      - Admin catalog
      operationId: listTheaters
      parameters:
      - name: cityId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseTheaterResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
      - basicAuth: []
    post:
      tags:
      - Admin catalog
      operationId: createTheater
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TheaterRequest'
            example:
              name: Central Cinema
              cityId: 00000000-0000-4000-8000-000000000001
              address: Baner, Pune
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TheaterResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
  /admin/api/v1/theaters/{theaterId}/auditoriums:
    get:
      tags:
      - Admin catalog
      operationId: listAuditoriums
      parameters:
      - name: theaterId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseAuditoriumResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
      - basicAuth: []
    post:
      tags:
      - Admin catalog
      operationId: createAuditorium
      parameters:
      - name: theaterId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditoriumRequest'
            example:
              name: Screen 1
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditoriumResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
...
  /admin/api/v1/movies:
    get:
      tags:
      - Admin catalog
      operationId: listMovies
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseMovieResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
      - basicAuth: []
    post:
      tags:
      - Admin catalog
      operationId: createMovie
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieRequest'
            example:
              durationMinutes: 120
              title: The Last Commit
              language: English
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
...
  /admin/api/v1/cities:
    get:
      tags:
      - Admin catalog
      operationId: listCities
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseCityResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
      - basicAuth: []
    post:
      tags:
      - Admin catalog
      operationId: createCity
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CityRequest'
            example:
              name: Pune
              timeZone: Asia/Kolkata
              country: India
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
  /admin/api/v1/auditoriums/{auditoriumId}/seats:
    get:
      tags:
      - Admin catalog
      operationId: listSeats
      parameters:
      - name: auditoriumId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseSeatResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
      - basicAuth: []
    post:
      tags:
      - Admin catalog
      operationId: createSeat
      parameters:
      - name: auditoriumId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeatRequest'
            example:
              rowLabel: A
              category: REGULAR
              seatNumber: 1
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
...
components:
  schemas:
...
    TheaterResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cityId:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
...
  responses:
    BadRequest:
      description: The request is malformed or fails validation.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Invalid request
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/invalid-request
            status: 400
            instance: /api/v1/seat-reservations
            code: INVALID_REQUEST
            detail: The request could not be processed.
    Unauthenticated:
      description: Valid HTTP Basic credentials are required.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Authentication required
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/authentication-required
            status: 401
            instance: /api/v1/seat-reservations
            code: AUTHENTICATION_REQUIRED
            detail: The request could not be processed.
    Forbidden:
      description: The authenticated account lacks the required role.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Access denied
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/access-denied
            status: 403
            instance: /api/v1/seat-reservations
            code: ACCESS_DENIED
            detail: The request could not be processed.
    NotFound:
      description: The resource does not exist or is concealed from this account.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Resource not found
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/resource-not-found
            status: 404
            instance: /api/v1/seat-reservations
            code: RESOURCE_NOT_FOUND
            detail: The request could not be processed.
...
    InternalError:
      description: An unexpected failure occurred.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Internal server error
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/internal-error
            status: 500
            instance: /api/v1/seat-reservations
            code: INTERNAL_ERROR
            detail: The request could not be processed.
  parameters:
...
    Page:
      name: page
      in: query
      description: Zero-based page index
      schema:
        type: integer
        format: int32
        default: 0
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic